tbl: proper width realization for colspanned cells
Commit 297f362 ("tbl: realize `w(...)` modifier for non-text blocks)
ignored the width of colspanned cells in almost all circumstances.

Consider a specification such as the following:

n s s
n r l

The cell in row 1, column 1 (n) is spanned and now has a distinct set of
left/right limits in comparison to row 2, column 1 (n).

In particular, after the widths of all non-colspanned "auto" and
"equalize" cells have been determined, we can account for the case of
row 1, column 1 being wider than the entire table width. The excess
width is distributed proportionally among the individual columns in that
column span.